use of unsigned int

+++ utils/random_seed.hpp
@@ -6,26 +6,26 @@
 #ifndef __UTILS__RANDOM_SEED__HPP__
 #define __UTILS__RANDOM_SEED__HPP__ 1
 
-#include <fstream>
 #include <ctime>
+#include <fstream>
 
 namespace utils
 {
   inline
-  unsigned long random_seed()
+  unsigned int random_seed()
   {
-    unsigned long seed;
+    unsigned int seed;
     
     std::ifstream udev("/dev/urandom");
     
     if (udev)
-      udev.read((char*) &seed, sizeof(unsigned long));
+      udev.read((char*) &seed, sizeof(unsigned int));
     
     if (udev.fail() || ! udev) {
       std::ifstream dev("/dev/random");
       
       if (dev)
-	dev.read((char*) &seed, sizeof(unsigned long));
+	dev.read((char*) &seed, sizeof(unsigned int));
       
       if (dev.fail() || ! dev)
 	seed = std::time(0);
